authorization

+++ routes/campgrounds.js
@@ -1,36 +1,11 @@
 const express = require('express');
 const router = express.Router();
 const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
-const { isLoggedIn } = require('../middleware');
+const { isLoggedIn, isAuthor, validateCampground } = require('../middleware');
 
 // To access/load database
 const db = require('../models/db_config');
 let sql;
-
-const validateCampground = (req, res, next) => {
-  if (!req.body.campground)
-    throw new ExpressError('Invalid Campground Data', 400);
-
-  const { title, location, image, price, description } = req.body.campground;
-  const priceNumber = Number(price);
-
-  if (!priceNumber) {
-    throw new ExpressError('Invalid Campground Data', 400);
-  } else if (priceNumber < 0) {
-    throw new ExpressError('Price must be greater than or equal to 0', 400);
-  } else if (!title) {
-    throw new ExpressError('Title must be required', 400);
-  } else if (!location) {
-    throw new ExpressError('Location must be required', 400);
-  } else if (!image) {
-    throw new ExpressError('Image must be required', 400);
-  } else if (!description) {
-    throw new ExpressError('Description must be required', 400);
-  } else {
-    next();
-  }
-};
 
 router.get(
   '/',
@@ -64,15 +39,16 @@ router.post(
   isLoggedIn,
   validateCampground,
   catchAsync(async (req, res, next) => {
-    sql = `INSERT INTO campgrounds (title, location, image, price, description)
-        VALUES (?, ?, ?, ?, ?)`;
+    sql = `INSERT INTO campgrounds (title, location, image, price, description, author)
+        VALUES (?, ?, ?, ?, ?, ?)`;
     const { title, location, image, price, description } = req.body.campground;
     const priceNumber = Number(price);
+    const author = req.user.id;
 
     const addedCampground = await new Promise((resolve, reject) => {
       db.run(
         sql,
-        [title, location, image, priceNumber, description],
+        [title, location, image, priceNumber, description, author],
         function (err) {
           if (err) {
             return reject(err);
@@ -93,7 +69,9 @@ router.get(
   '/:id',
   catchAsync(async (req, res) => {
     const campId = req.params.id;
-    sql = `SELECT * FROM campgrounds WHERE id = ?`;
+    sql = `SELECT campgrounds.*, users.username, users.email FROM campgrounds 
+      INNER JOIN users ON users.id = campgrounds.author
+      WHERE campgrounds.id = ?`;
 
     const campground = await new Promise((resolve, reject) => {
       db.get(sql, [campId], (err, rows) => {
@@ -104,7 +82,8 @@ router.get(
       });
     });
 
-    sql = `SELECT * FROM reviews WHERE camp_id = ?`;
+    sql = `SELECT reviews.*, users.username, users.email FROM reviews 
+      INNER JOIN users ON users.id = reviews.author WHERE reviews.camp_id = ?`;
     const reviews = await new Promise((resolve, reject) => {
       db.all(sql, [campId], (err, rows) => {
         if (err) {
@@ -127,25 +106,20 @@ router.get(
 router.get(
   '/:id/edit',
   isLoggedIn,
+  isAuthor,
   catchAsync(async (req, res) => {
     const campId = req.params.id;
-    sql = `SELECT * FROM campgrounds WHERE id=?`;
+    sql = `SELECT * FROM campgrounds WHERE id = ?`;
     const campground = await new Promise((resolve, reject) => {
       db.get(sql, [campId], (err, rows) => {
         if (err) {
           return reject(err);
         }
-
         resolve(rows);
       });
     });
 
-    if (!campground) {
-      req.flash('error', 'Cannot find that campground!');
-      res.redirect('/campgrounds');
-    } else {
-      res.render('campgrounds/edit', { campground });
-    }
+    res.render('campgrounds/edit', { campground });
   })
 );
 
@@ -153,19 +127,21 @@ router.get(
 router.put(
   '/:id',
   isLoggedIn,
+  isAuthor,
   validateCampground,
   catchAsync(async (req, res) => {
     const campId = req.params.id;
+    const author = req.user.id;
     const { title, location, image, price, description } = req.body.campground;
     const priceNumber = Number(price);
 
     sql = `UPDATE campgrounds SET title = ?, location = ?, image = ?, price = ?, description = ? 
-          WHERE id = ?`;
+          WHERE id = ? AND author = ?`;
 
     const editedCampground = await new Promise((resolve, reject) => {
       db.run(
         sql,
-        [title, location, image, priceNumber, description, campId],
+        [title, location, image, priceNumber, description, campId, author],
         function (err) {
           if (err) {
             return reject(err);
@@ -176,10 +152,6 @@ router.put(
       );
     });
 
-    if (editedCampground.changes === 0) {
-      throw new ExpressError('Campground not found', 404);
-    }
-
     req.flash('success', 'Succesfully updated campground!');
     res.redirect(`/campgrounds/${campId}`);
   })
@@ -189,6 +161,7 @@ router.put(
 router.delete(
   '/:id',
   isLoggedIn,
+  isAuthor,
   catchAsync(async (req, res) => {
     const campId = req.params.id;
 
@@ -214,17 +187,14 @@ router.delete(
             resolve({ changes: this.changes });
           });
         });
-
-        if (deletedReview.changes === 0) {
-          throw new ExpressError('Review not found', 404);
-        }
       });
     }
 
-    sql = `DELETE FROM campgrounds WHERE id = ?`;
+    const author = req.user.id;
+    sql = `DELETE FROM campgrounds WHERE id = ? AND author = ?`;
 
     const deletedCampground = await new Promise((resolve, reject) => {
-      db.run(sql, [campId], function (err) {
+      db.run(sql, [campId, author], function (err) {
         if (err) {
           return reject(err);
         }
